Refactor: 코드분리
main.py내용을 schema.py, services.py, main.py로 나눔
